feat: Implement static affiliate permission handling
Introduces a new `staticAffiliatePermission` utility to manage access for affiliate-specific routes.

The `getOutsideRoutes` function now accepts an `isAffiliate` flag. When `isAffiliate` is true, the `staticAffiliatePermission` function is used to determine authorization, overriding the default `isSuperAdmin` or `hasCategoryPermission` checks. This ensures that affiliate users are granted access based on their specific roles and permissions defined in `staticAffiliatePermission.js`.

The `App.jsx` component has been updated to pass `true` for the `isAffiliate` flag when rendering `affiliateOutsideRoute` to correctly apply the new permission logic.

Additionally, the `BASE_URL` in `ApiList.js` has been switched back to the production URL.